+++ src/mantle/http/routing/class-route.php
@@ -28,8 +28,11 @@ use function Mantle\Support\Helpers\get_callable_fqn;
 
 /**
  * Route Class
+ *
+ * Used in both HTTP routing and WordPress REST API routing.
  */
 class Route extends Symfony_Route {
+	use Concerns\Registers_Rest_Route;
 	use Route_Dependency_Resolver;
 
 	/**
@@ -76,7 +79,7 @@ class Route extends Symfony_Route {
 	 * @param string                       $path The path the route responds to.
 	 * @param \Closure|array<mixed>|string $action The route callback or array of actions.
 	 */
-	public function __construct( array $methods, string $path, $action ) {
+	public function __construct( array $methods, string $path, \Closure|array|string $action ) {
 		parent::__construct( $path );
 
 		$this->setOption( 'utf8', true );
@@ -91,8 +94,10 @@ class Route extends Symfony_Route {
 			];
 		} elseif (
 			is_array( $action ) // @phpstan-ignore-line function.alreadyNarrowedType
-			&& ! empty( $action[0] )
-			&& ! empty( $action[1] )
+			&& array_is_list( $action )
+			&& count( $action ) === 2
+			&& isset( $action[0] )
+			&& isset( $action[1] )
 			&& is_string( $action[0] )
 			&& is_string( $action[1] )
 			&& class_exists( $action[0] )
@@ -100,9 +105,9 @@ class Route extends Symfony_Route {
 			/**
 			 * Handle controller 'static' style callbacks.
 			 *
-			 * They're written as callable-style (class name -> method). For PHP 8,
-			 * they need to be manually detected since is_callable() will return false
-			 * for non-static methods using the array structure.
+			 * They're written as callable-style (class name -> method). They need to
+			 * be manually detected since is_callable() will return false for
+			 * non-static methods using the array structure.
 			 */
 			$action = [
 				'callback' => $action,