+++ tests/Http/Routing/RestApiRoutingTest.php
@@ -4,6 +4,7 @@ namespace Mantle\Tests\Http\Routing;
 use Closure;
 use InvalidArgumentException;
 use Mantle\Facade\Route;
+use Mantle\Http\Routing\Route as RouteObject;
 use Mantle\Http\Controller;
 use Mantle\Testing\Concerns\Refresh_Database;
 use Mantle\Testing\FrameworkTestCase;
@@ -12,22 +13,60 @@ use WP_REST_Request;
 class RestApiRoutingTest extends FrameworkTestCase {
 	use Refresh_Database;
 
-	public function test_generic_route() {
-		Route::rest_api(
+	public function test_register_rest_with_callback_directly(): void {
+		$route = Route::rest_api(
 			'namespace/v1',
-			'/example-closure-third',
-			fn () => 'example-closure-third',
+			'/example-closure',
+			fn () => 'example-closure',
 		);
 
-		Route::rest_api(
+		$this->assertInstanceOf( RouteObject::class, $route );
+
+		$this->get( rest_url( '/namespace/v1/example-closure' ) )
+			->assertOk()
+			->assertContent( json_encode( 'example-closure' ) );
+
+		$this->post( rest_url( '/namespace/v1/example-closure' ) )
+			->assertNotFound()
+			->assertJsonPath( 'code', 'rest_no_route' );
+	}
+
+	public function test_register_rest_with_callback_in_argument_array(): void {
+		$route = Route::rest_api(
 			'namespace/v1',
-			'/example-array-third',
+			'/example-array',
 			[
-				'callback' => fn () => 'example-array-third',
+				'callback' => fn () => 'example-array',
 			]
 		);
 
-		Route::rest_api(
+		$this->assertInstanceOf( RouteObject::class, $route );
+
+		$this->get( rest_url( '/namespace/v1/example-array' ) )
+			->assertOk()
+			->assertContent( json_encode( 'example-array' ) );
+
+		$this->post( rest_url( '/namespace/v1/example-array' ) )
+			->assertNotFound()
+			->assertJsonPath( 'code', 'rest_no_route' );
+	}
+
+	public function test_register_with_string_callback(): void {
+		$route = Route::rest_api(
+			'namespace/v1',
+			'/example-string',
+			__NAMESPACE__ . '\testable_function_name',
+		);
+
+		$this->assertInstanceOf( RouteObject::class, $route );
+
+		$this->get( rest_url( '/namespace/v1/example-string' ) )
+			->assertOk()
+			->assertContent( json_encode( 'function-response' ) );
+	}
+
+	public function test_register_routes_in_callback(): void {
+		$return = Route::rest_api(
 			'namespace/v1',
 			function() {
 				Route::get( '/example-group-get', fn () => 'example-group-get' );
@@ -41,29 +80,7 @@ class RestApiRoutingTest extends FrameworkTestCase {
 			}
 		);
 
-		Route::rest_api(
-			'namespace/v1',
-			function () {
-				Route::get( '/example-invoke', Testable_Invokable_Rest_Api_Controller::class );
-
-				Route::get( '/example-controller/index', [ Testable_Rest_Api_Controller::class, 'index' ] );
-				Route::get( '/example-controller/show', [ Testable_Rest_Api_Controller::class, 'show' ] );
-			},
-		);
-
-		Route::rest_api(
-			'namespace/v1',
-			'/example-string-function',
-			__NAMESPACE__ . '\testable_function_name',
-		);
-
-		$this->get( rest_url( '/namespace/v1/example-closure-third' ) )
-			->assertOk()
-			->assertContent( json_encode( 'example-closure-third' ) );
-
-		$this->get( rest_url( '/namespace/v1/example-array-third' ) )
-			->assertOk()
-			->assertContent( json_encode( 'example-array-third' ) );
+		$this->assertNull( $return );
 
 		$this->get( rest_url( '/namespace/v1/example-group-get' ) )
 			->assertOk()
@@ -76,6 +93,19 @@ class RestApiRoutingTest extends FrameworkTestCase {
 		$this->post( rest_url( '/namespace/v1/example-post' ) )
 			->assertOk()
 			->assertContent( json_encode( 'example-post' ) );
+	}
+
+	public function test_controller_route() {
+		// Registering a group of routes from controller methods.
+		Route::rest_api(
+			'namespace/v1',
+			function () {
+				Route::get( '/example-invoke', Testable_Invokable_Rest_Api_Controller::class );
+
+				Route::get( '/example-controller/index', [ Testable_Rest_Api_Controller::class, 'index' ] );
+				Route::get( '/example-controller/show', [ Testable_Rest_Api_Controller::class, 'show' ] );
+			},
+		);
 
 		$this->get( rest_url( '/namespace/v1/example-invoke' ) )
 			->assertOk()
@@ -88,13 +118,9 @@ class RestApiRoutingTest extends FrameworkTestCase {
 		$this->get( rest_url( '/namespace/v1/example-controller/show' ) )
 			->assertOk()
 			->assertContent( json_encode( 'show-response' ) );
-
-		$this->get( rest_url( '/namespace/v1/example-string-function' ) )
-			->assertOk()
-			->assertContent( json_encode( 'function-response' ) );
 	}
 
-	public function test_middleware_route() {
+	public function test_middleware_class_route() {
 		Route::middleware( Testable_Before_Middleware::class )
 			->rest_api(
 				'namespace/v1',
@@ -104,13 +130,18 @@ class RestApiRoutingTest extends FrameworkTestCase {
 				}
 			);
 
+		$this->get( rest_url( '/namespace/v1/example-middleware-route' ) )
+			->assertOk()
+			->assertContent( json_encode( 'middleware-response' ) );
+	}
+
+	public function test_middleware_closure_route(): void {
 		Route::middleware(
-			function( WP_REST_Request $request, $next ) {
+			function( WP_REST_Request $request, Closure $next ) {
 				$request->set_param( 'input', 'modified' );
 				return $next( $request );
 			}
-		)
-		->rest_api(
+		)->rest_api(
 			'namespace/v1',
 			'/example-middleware-modify-post',
 			[
@@ -119,13 +150,26 @@ class RestApiRoutingTest extends FrameworkTestCase {
 			]
 		);
 
-		$this->get( rest_url( '/namespace/v1/example-middleware-route' ) )
-			->assertOk()
-			->assertContent( json_encode( 'middleware-response' ) );
-
 		$this->post( rest_url( '/namespace/v1/example-middleware-modify-post' ), [ 'input' => 'value' ] )
 			->assertOk()
 			->assertContent( json_encode( 'modified' ) );
+	}
+
+	public function test_middleware_added_fluently(): void {
+		Route::rest_api(
+				'namespace/v1',
+				function (): void {
+					Route::get( '/example-middleware-fluent', function() {
+						return 'base-response';
+					} )->middleware( function ( WP_REST_Request $request, Closure $next ) {
+						return 'middleware-response';
+					} );
+				},
+		);
+
+		$this->get( rest_url( '/namespace/v1/example-middleware-fluent' ) )
+			->assertOk()
+			->assertContent( json_encode( 'middleware-response' ) );
 	}
 
 	public function test_group_route() {
@@ -144,16 +188,6 @@ class RestApiRoutingTest extends FrameworkTestCase {
 		$this->get( rest_url( '/namespace/v1/example-group' ) )
 			->assertOk()
 			->assertContent( json_encode( 'middleware-response' ) );
-	}
-
-	public function test_invalid_action() {
-		$this->expectException( InvalidArgumentException::class );
-
-		Route::rest_api(
-			'namespace/v1',
-			'/example-invalid-action',
-			'invalid-action',
-		);
 	}
 }